Remodel translation algorithm
The previous version didn't cover most of the corner cases and
was quite messy, with too many nested iterations, trying to build
the final string in 'one shot'. The current version parses the
integer and builds a list of strings for each translated digit,
handling special cases at and filtering out zeos at the end.

Signed-off-by: Jonatas Rech <jonatas.rech@gmail.com>